refactor: [#31] unit tests for workflow actions
I want to test eac workflow step independently. The only injected
dependency was Typer. If we remove that dependency we can easily test
the isolated action. This is a test but I think it's a good idea becuase
tests are already slow. We have to run a lot of tests for each corner
case in the workflow, and for each test we have the build the initial
state. It would take too much time.